Coverage Report
Please consider this coverage report, as it test the latest version of the code. We want to mention that the test for the configuration.py module does not satisfy the requirement of >75% coverage,  because there was a problem when we upload in Actions Github, so we have to comment the test_read in test_configuration.py file. We tried several times in our local machine and the unit test for this method was working but for any reason it wasn't when we upload in GitHub Actions.